Disable reliable sender
I'll be replacing reliability with redundancy; basically, instead of requiring acks for reliable packets, both the client and server will stack and send several previous states redundantly. This has several benefits, such as allowing the frequency at which packets are sent to be greatly reduced (at the cost of larger packets, but that's a non-issue really, the overhead of many tiny packets is probably worse), and it makes implementing the authoritative server a lot easier.

The reliable sender will probably be needed later to make clients able to download subs again, but for now it's not important.