Started overhauling the campaign. The goal is to give the player a greater sense of progression and exploration, make the world feel more alive, create a basis for the eventual in-game stations/outposts/etc and add a more clear end goal.
Now the campaign map consists of 9 concentric rings. The end goal is to reach the center, with difficulty increasing as the player progresses closer to it (todo: make difficulty have a proper effect in-game). Initially colonies and inhabited outposts only exist at the outermost rings, but the plan is that as the player progresses towards the center and clears the passageways and uninhabited locations of monsters, new colonies will gradually start popping up closer to the center. There's also the chance that some of these new colonies won't make it if the player doesn't support them by helping them deal with monster attacks and such.

Some other changes:
- Mission prefabs.
- CharacterInfo caches the character config files. Reloading the human config file every time when generating hireable characters for locations took an excessive amount of time.
- ToolBox.GetRandomLine caches the lines in a file. Generating the names for the map locations and hireable characters was also starting to take up way too much time.