WIP "event director" and ScriptedEvent refactoring.
Instead of creating all events at the start of the round, EventManager now monitors the "intensity" of the situation and attempts to create events at moments when there's not much going on. The goal is to improve the pacing of the game by reducing the probability of random events in crisis situations and increasing the probability when things get more quiet.

The "intensity" is calculated based on the health of the crew, the amound and dangerousness of nearby enemies and the amount of hull breaches, fires and water in the sub. Additional factors and tweaking their weights is most likely needed to make the estimation more accurate and to prevent abusing the system. If the intensity gets below a specific threshold and it's been long enough since the last event, the EventManager creates a new random event. The threshold also increases over time (resetting with each new event), so there will be new events eventually even if a crisis situation is not resolved.

EventManager is also fully server-side now, meaning that servers can create events any way they wish without the need for client-side modifications.