Fixes to server crashing.
InGameInfo was moved to GameSession for singleplayer and client compatability but it was not initiated before being used in multiplayer to record clients resulting in null reference exceptions before players were added to the netlobby playerlist, This should resolve it.

Additionally Modified Project Items to stop applying the content folder and configs automatically when running debug/release.