New health system (WIP):
Damage is applied individually to each limb and there are three types of damage: blunt, bleeding and burn. Characters don't have an individual "health" value anymore, but instead "vitality" which represents the overall condition of the character. Vitality can be reduced by blunt and burn damage, lack of oxygen, bloodloss, poisons, sicknesses (NYI) and loss of mental health (NYI). Characters fall unconscious if vitality decreases below zero, and die if it drops to -100.