Railgun Shell Rack
Adds a rack capable of holding 3 Railgun shells,
so players / sub makers won't have to store shells on the floor anymore. Also helps the crew keep track of how many Railgun shells they sub have. 

Because the racks take up space, it also would help with balancing subs for PvP and so on.

texture located in "content/items/weapons/railgunetc2.png", which it shares with the new loaders and rear facing controller.
https://imgur.com/muAIF9P

This feature has been requested a lot lol, so I hope this will make a lot of people really happy