Fixed face attachments not matching between server and clients.
If the client sent an invalid attachment index (which they always do atm, because the indices are not set by the clients and have the default value of -1), the server would simply send the invalid value (255) to clients and cause them to randomize the attachments.

The attachments of the host's character didn't match because the latter part of the CharacterInfo constructor was skipped if generating a random name is not needed, causing LoadHeadAttachments not to be called at all, so the character would not have any attachments server-side. The clients would however randomize the host's attachments because of the invalid attachment indices.