Cleaner Penumbra texture, more pointlight variations
- Removed the AmbientLightIntensity variable (Forgot level editor already has a way to edit it).
- Cleaned up Penumbra texture and made it higher resolution.
- Added more pointlight variations to test.

pointlight_bright.png could be used as a replacement for the old light texture.

pointlight_bounce.png the currently in use one, has a outer light ring which acts as a bounce light, when light bounces from the wall opposite of the light source filling the room with light.

pointlight_falloff.png is the same as the bounce variation just without the outer light ring.