Changes to FishAnimController logic (similar to the changes in 244bdfc). Instead of attempting to rotate the limbs in the "shortest" direction towards the desired direction, the rotation of the main limb is taken into account to prevent rotating limbs in the "wrong direction".
Also had to change how rotation widgets work in the character editor: they now go from -180 to 180 instead of 0 to 360. Required for the new logic to work, for example if we want to rotate the head -10 degrees relative to the torso, we can't use 350 anymore because it'd cause the head joint to actually bend 350 degrees instead of being interpreted as -10.

I also had to do a bunch of changes to the creature ragdolls/animations to get the working correctly again, but tried to keep the animations close to how they were. They seem to be working much more reliably now, but I still need to do some more testing.