Display the controlled character in the crewmanager interface. Hiding the controlled character makes the interface a little more confusing to use because the order/positions of the characters changes depending on which character is being controlled.
It's also now possible to give orders to the character you're controlling. In single player it can be useful if you want the controlled character to keep doing something when switching to another one, in the multiplayer it can be used to let others know what you're doing, and in both SP and MP it can be used to enable the order indicators (arrows pointing towards the reactor, weapons, etc).

Closes #1101