Experimental changes to item position syncing. Previously the items only had a boolean "NeedsPositionUpdate" property that got flagged if the item had moved enough from the last position sent to clients, which didn't work that well: if the clients missed the message or if the item moved somewhere else client-side, the position wouldn't get corrected until the item moves again server-side.
Now the items have a "PositionUpdateInterval" field which gradually increases and gets reset when the item moves (faster movement = more frequent updates). As a result, the server keeps sends frequent updates for moving items and gradually slows down the update rate when the item comes to rest. Updates are also sent for non-moving items, just very infrequently, which should ensure that the positions will eventually get corrected even if the item somehow ends up in an incorrect position client-side. The distance between the items and the clients also effects the update rate: far-away items are updated less frequently (or not at all if the item is very far).