Campaign fix: clear missions from locations that change their type, and all adjacent locations. Caused missions to still be available when they logically shouldn't be (e.g. a transport mission from an uninhabited location to another) and syncing issues in multiplayer.
The clients would access the AvailableMissions property as soon as the campaign map is received, generating a bunch of missions for the initial location type, and later when they receive the up-to-date save they'd (maybe) generate some more. The server didn't access it until someone selects a location, potentially causing a mismatch in the available missions, which could in the worst case cause complete desync (for example if the server spawns an artifact while the clients don't). Could be related to some of the issues mentioned in #1340