Revert 8e2e2526 ("Spawn multiple monsters using the same spatial and temporal spreading logic as in events". Caused entity IDs to often differ between the server and clients in monster missions, leading to all sorts of syncing issues (the most frequent being IndexOutOfRange exceptions when attempting to read data for an incorrect entity).
Since the monster spawning logic is run by both the server and the clients, there's a chance the server spawns additional entities before the monster, and the clients won't know about them until after the monster spawns, causing them to assign the monsters different IDs. The RNG calls can also cause problems, because the RNG methods must be called in the exact same order for them to produce the same values (which is probably not the case when using them at a random time during the round). TODO: make the monster missions server-side like monster events, only reporting the IDs of the target monster(s) to the clients.