Fixed another entity ID mismatch issue that caused "missing entity" errors in multiplayer. NetEntityEventManager would write the _current_ ID of the entity in the message, even though the client does not necessarily know about the ID being changed yet.
For example, if the game created an event for an alien pump with the id 123, then loaded the player's character who happened to have a headset with the ID 123, the pump would get assigned a new ID and EntitySpawner would create an event that lets the client know about the ID change. But when the server wrote the event for the pump, it used the new ID, which the client would not know about because the EntitySpawner event had not been received yet.