Improved the logic used in Steering to determine which side a docking port needs to be relative to another one to dock. Closes #1747
Previously the direction was determined based on the position of the port relative to the sub's origin (e.g. a port above the sub's center was considered a port that can only be docked to from above), which caused issues in more exotic shaped subs (for example ones with a "shuttle slot" partially inside the sub).