Use Lidgren's NetBitWriter to Read/Write network messages instead of reimplementing the logic ourselves.
The methods now behave like Lidgren's Read/Write methods, and do not automatically pad the values with extra bits. The automatic padding caused a bunch of issues because padding may be needed when writing the data to the original outgoing message, but not when reading the data from a buffer extracted from the message (for example in Item.DoDelayedCorrection or when clients extract server settings from a lobby update message), or vice versa.