Fixed traitor missions almost always placing the items inside the same containers.
The previous logic would work if the item list was in a completely random order, but it isn't: the items in the sub are generally close to each other in the list, with hundreds/thousands of non-sub items after them (ruin items, level resources, items in outposts). The randomly selected index usually tends to be somewhere outside the range of the container indices, and iterating through the list will cause the first suitable container in the list to be picked.