Extract player details dialog out of NetLobbyScreen.
Multiple locations were calling the SelectPlayer function of NetLobbyScreen.
This function created a details dialog for the given player client. Since
this is used outside of NetLobbyScreen and has no specific ties to it, it
should be a separate class. Created a PopUpDialog class which tracks active
dialogs for activation purposes, and is intended to act as a root for dialogs
that block out the background. The SelectPlayer function has been refactored
into PlayerDetailsDialog with a focus on making the code more readable by
splitting it into functions handling the different sections.

Expected behavioral changes:
- Previously clicking the background behind the dialog did not close the
  dialog due to elements being out of order, this has been fixed so that it
  closes the dialog.
- When checking/unchecking a permission, if it matches a preset then the rank
  dropdown will be set to that preset.