WIP: Use ternary states for some of the server filters.
This is a preliminary attempt at making some of the server filters like Karma
and Traitors use ternary options (All/Enabled/Disabled) instead of a tickbox
which only gives you two of the states. I went with the approach of using drop
down boxes instead of trying to implement a third state into GUITickBox. I
couldn't think of a good way to graphically represent the third state, thus
the usage of drop downs.

This change is functional, but there are some spots where it needs attention.
I left TODO comments where I was unsure of how to proceed. This could be used
as a jumping-off point for the devs to implement this feature fully if they
like the approach.

Translation requirements:
- Needs a new 'servertag.X.label' for the tags affected here. For example,
  "Traitors".
- Needs 'servertag.X.true' to be modified to not include the label. For example,
  "Traitors enabled" -> "Enabled"
- Needs 'servertag.X.false' to be modified to not include the label. For example,
  "Traitors disabled" -> "Disabled"

Fixes #3652