Allows the husk infection to carry player control over to the husk.
The husk infection has two new parameters: "ControlHusk" and "ControlException".
ControlHusk decides whether or not the player should be allowed to control their husk.
ControlException will prevent the player from controlling their husk if their species is in this list.

Added two console commands: "ignorehuskexceptions" and "restorehuskexceptions". These commands will toggle a new boolean parameter of the husk infection which lets it ignore any exception lists if set to true.

Added a new function for characters that returns the name of the player controlling them when called.