Added ReviveCharacter effect to status effects
Added a new variable to clients, "DeadCharacter". Initally null, gets set to the character you were controlling prior to dying on death.

Added a new usable effect to status effects, "ReviveCharacter", which also accepts an attribute "noseveredlimbs" (false by default).

When using the "ReviveCharacter" effect, the targeted character will be revived and whoever was controlling it prior to dying will be set in control of the character again. If "noseveredlimbs" is set to true, the character will only be revivable if it has all of its limbs intact.