Added damage memory system for enemies
Enemies can now remember who damaged them and how much damage was dealt to them by a specific character, and can also remember damage dealt to friendly characters nearby. Damage memories eventually fade away.

Targeting parameters now accept a new attribute, "MinDamageToIgnoreAfflictionTag".

If a character has an affliction tag which changes the behavior of an enemy (for example, change to idle state), this state will now be ignored by the AI if the character deals enough damage to it (depending on the value of "MinDamageToIgnoreAfflictionTag"), which will then go back to using the default state defined in the targeting parameters to deal with that character.