Allow to define character types for jobs
Job prefabs now accept a new element: "charactertype".
By setting up the subelements of this new element: "identifier" and "spawnoutside", it is now possible to make it so certain jobs will make the client spawn and respawn as a character other than a human (inside or outside the submarine, depending on the "spawnoutside" boolean) in any gamemode other than campaign.

Multiple character types can be set up in order to make the client spawn as different characters each time they are spawned / respawned.

If the character has the same inventory slots as a human, it will automatically equip the items that the human would have gotten with that job.