Allow relateditem to check for items in a specific character inventory slot
The relateditem conditional now accepts two new parameters:
"checkcharacterinventoryslot": a boolean which toggles whether or not the item should only be looked for in a specific inventory slot.
"characterinventoryslottype": a (case sensitive) string which should be set to one of the inventory slot types, which is the slot that will be targeted by the conditional.