Allow characters to switch attacks without snapping ropes from previous attacks
Adds a new parameter to attacks: "SnapRopeOnNewAttack", set to true by default.
If true, when a new attack is picked to be used next by a character, any rope created by a previous attack will automatically snap.
If false, when a new attack is picked to be used next by a character, ropes will be left intact.