Fix power indicator not rotating with battery.
PR for issue #9848

Drawing the charge indicator did not take into account the rotation
of the parent container. This fixes this by adding two functions for
drawing rotated rectangles, one for an outline and one for filled.
The function for drawing the outlined rectangle has been written to
accept the desired position for the outline for increased ease of
use.

Behavior maintained from previous implementation:
- Mirroring the parent X or Y does not mirror the indicator.

Behavior changed/added:
- The outline is now drawn on the inside of the given coords, so that
  the overall drawn size matches the size provided.
- The charge indicator accounts for the thickness of the outline so
  that very small charge levels are still visible and not covered by
  the outline.
- Added a 'FlipIndicator' property to the power container so that it
  is possible to truely mirror the object if desired.