Updated entire fileset
As you requested and thanks to yours help - we added a hint that tells the player about grenade's new functionality as soon as they pick one up.
To make it possible we also added a new tag to all throwable grenades - "throwableexplosive".

As for the ways to further resolve communication issue, grenades now have particles of pin and lever falling from them even if they are thrown without premature cooking, which might suggest to the player that said safety mechanisms can be disengaged manually. From there i think it would be logical to rename the "Use" key binding in settings to "Activate", as in game it is only use to activate flares and, if this rework gets accepted, arm grenades

As to why we uploaded entire fileset including pictures and sounds instead of just needed changes - we found out that atleast why downloading them from github, they became corrupted, so we did this replacement as purely safety measure

Hope you are satisfied with our work now!