Backwards compatible engine thrust for new drag formula
A sub 10 times smaller used to need 10 times less thrust force to
achieve the same speed. The new drag formula needs about 4.6 times more
than that. This change in the thust-to-weight ratio means faster
acceleration for drones and shuttles.
Calculate this factor to prevent the top speed from changing in existing
submarine designs.