[fix] Use filepath instead of XDocument for md5hash (submarine)
Prior to this commit, it would load an XML file (ignoring white spaces),
write XDocument into a string (Loading the game would peak at 150MB just
for this function alone). It would then try to strip whitespaces again
(which allocates another 50-70mb as it constructs a string builder with
half the current string size) and finally hash it.

Instead, read the FilePath (which could be cached by the OS already),
strip whitespaces and hash.

With this change, this function no longer allocates enough to appear in
DPA.